+++ src/pymdgen/test_module.py
@@ -2,6 +2,8 @@
 A module to use as a target during unit tests
 """
 
+from functools import wraps
+
 
 class prop(object):
     help = "property help"
@@ -44,3 +46,17 @@ class dummy(object):
 def dummy_func():
     """ this is a dummy func """
     return
+
+def decorate_this(fn):
+    @wraps(fn)
+    def wrapped(*args):
+        return fn(*args)
+
+    return wrapped
+
+@decorate_this
+def decorated_func(specifc):
+    """ this is a decorated function """
+    return specific
+
+